Refactor post-analysis scripts and remove unused Python script
Improved comments and English translations in analisi_indels_NNN.sh. Refactored post_analisi_indel.sh and post_analisi_snp.sh to streamline chromosome-specific target extraction and header handling, replacing awk with grep for efficiency. Removed the unused convert_gnomAD.py script. Updated submit_job_automated_new_multiple_vcfs.sh to adjust CI/CD test handling and conditional directory cleanup.

+++ PostProcess/submit_job_automated_new_multiple_vcfs.sh
@@ -433,10 +433,6 @@ while read vcf_f; do
 	fi
 	# END STEP 3 - off-targets search
 
-	if [[ "$cicd_test" == "True" ]]; then  # if CI/CD test stop execution here
-		exit 0
-	fi
-
 	# START STEP 4 - off-targets post-analysis
 	cd "$starting_dir"
 	echo -e "Start post-analysis"
@@ -755,7 +751,9 @@ echo -e "Cleaning directory"
 rm -f *.CFDGraph.txt
 rm -f indels.CFDGraph.txt
 rm -r "crispritz_prof"
-rm -r "crispritz_targets" # remove targets in online version to avoid memory saturation
+if [[ "$cicd_test" != "True" ]]; then  # if running complete-test do not rempve for validation
+	rm -r "crispritz_targets" # remove targets in online version to avoid memory saturation
+fi
 # change primary and alt filenames
 mv $final_res "${output_folder}/$(basename ${output_folder}).bestMerge.txt"
 mv $final_res_alt "${output_folder}/$(basename ${output_folder}).altMerge.txt"
